update env, sign, proteksi app

+++ frontend/src/api/api.ts
@@ -1,44 +1,32 @@
-import axios, {
-  AxiosError,
-  AxiosResponse,
-  InternalAxiosRequestConfig,
-} from "axios";
+import axios, { AxiosError, AxiosResponse, InternalAxiosRequestConfig } from "axios";
 
 const API = axios.create({
   baseURL: import.meta.env.VITE_API_URL,
 });
 
-/* =======================
-   REQUEST INTERCEPTOR
-======================= */
 API.interceptors.request.use(
   (config: InternalAxiosRequestConfig) => {
     const token = localStorage.getItem("token");
-
     if (token) {
-      config.headers.Authorization = token; // tanpa Bearer
+      config.headers.Authorization = `Bearer ${token}`; // HARUS ada "Bearer "
     }
-
     return config;
   },
   (error: AxiosError) => Promise.reject(error)
 );
 
-/* =======================
-   RESPONSE INTERCEPTOR
-======================= */
 API.interceptors.response.use(
   (response: AxiosResponse) => response,
   (error: AxiosError) => {
     if (error.response) {
       const status = error.response.status;
+      const url = error.config?.url;
 
-      if (status === 401 || status === 403) {
+     if ((status === 401 || status === 403) && !url?.includes("/auth/login")) {
         localStorage.clear();
-        window.location.href = "/signin";
+        window.location.href = "/signin"; 
       }
     }
-
     return Promise.reject(error);
   }
 );
